Created models folder, move the ICategory interface to the models folder, fix all the imports
